Merged PR 18828: aroundPriceFinderPlugin
## What's being changed

We are adding a plugin around getMinPrices().

## Why it's being changed

When syncing a shared catalog, we need to source 'tier' prices as set for products in the relevant shared catalog.

This PR is suggested as an alternative to the initial di.xml / preference approach. I believe this is simpler. Overriding getMinPrices doesn't work so well with di preference, because Email and B2b are dependent modules, therefore di.xml is combined, and the B2b version of `PriceFinder` is always used.

I'm not completely happy with having to hand a `$module` flag to the exporter, but can't see a way around it.

## How to review / test this change

This PR should be tested by checking out in this branch https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension/pullrequest/18789?_a=files in Email Module

- Set up a shared catalog and assign in to this product of all types (simple products , configurable, grouped bundled)
- assign tier prices for all
- leave at least one simple product without tier price
- run sync catalog (increase batch size to 500, you may need to run multiple times catalog sync to get all shared catalog items imported)
- Make sure prices for grouped  , configurable and bundled products are 0
- Make sure you get the tier price for simple product that you assigned a discount
- Make sure you get the normal price for the simple product that you didn't assigned a tier price

Related work items: #105982