Merged PR 34456: Handle company not found in contact sync
## What's being changed

The B2B CustomerPlugin.

## Why it's being changed

We have a method that fetches a company's store credit balance, for possible use in the STORE_CREDIT_BALANCE data field. If the supplied company ID is not found in Magento, it throws a `NoSuchEntityException`, which derails the contact sync.

This PR try/catches the credit balance fetch.

## How to review / test this change

- Enable the B2B module
- Create a customer attached to a company
- In `Dotdigitalgroup\B2b\Helper\Data` set `$company->getId()` on L80 with a fake id like 3089
- Run contact sync
- Note the log line generated with the 'Requested company not found' error exception
- Confirm the contact sync completes

## Notes

The underlying issue will occur regardless of whether or not STORE_CREDIT_BALANCE is mapped.

Related work items: #154512